+++ core/unit_test/incremental/Test12b_TeamScratch.hpp
@@ -79,7 +79,7 @@ struct TeamScratch {
     auto v_H = Kokkos::create_mirror_view_and_copy(Kokkos::HostSpace(), v);
 
     size_t check   = 0;
-    const size_t s = pN * sX * sY;
+    const size_t s = static_cast<size_t>(pN) * sX * sY;
     for (int n = 0; n < pN; ++n)
       for (int m = 0; m < sX; ++m) check += v_H(n, m);
     ASSERT_EQ(check, s * (s - 1) / 2);
